Adding push notifications
The two services added in this commit handle registering for a
Firebase InstanceId (which will not change throughout the app lifecycle) and a
token, which will change throughout the lifecycle of the app (every few months or so).
We'll need to post that to the server whenever the master branch of the server project starts
supporting that feature.

The second service (JLMFirebaseMessagingService) actually handles receiving the messages.
Currently it just throws a notification up (whether in the background or in the foreground),
but later on, we'll have it be a bit smarter.

Both services are defined in the AndroidManifest, where we also define the intent-filters
we want them to respond to (Firebase intent-filters in our case)